requirements.txt
wxauto
google-generativeai>=0.5.0
Pillow